requirements.txt
asgiref==3.6.0
autopep8==2.0.1
Django==4.1.6
django-cors-headers==3.13.0
django_debug_toolbar==3.8.1
djangorestframework==3.14.0
gunicorn==20.1.0
mysqlclient==2.1.1
Pillow==9.4.0
pycodestyle==2.10.0
pytz==2022.7.1
sqlparse==0.4.3
tomli==2.0.1
whitenoise==6.3.0
